requirements.txt
huggingface_hub==0.23.2
flask==3.1.0
opencv-python-headless==4.10.0.84
pinecone-client==5.0.1
pinecone[grpc]
pinecone==5.4.2
pydub==0.25.1
ffmpeg-python==0.2.0
python-ffmpeg==2.0.12
deepface==0.0.93
facenet-pytorch==2.6.0
opencv-python-headless==4.10.0.84
tensorflow==2.18.0
protobuf==5.28.3
tf-keras==2.18.0
python-dotenv==1.0.1
boto3==1.35.83
ipython==8.30.0